+++ quantify_scheduler/backends/zhinst/__init__.py
@@ -1,0 +1,3 @@
+from quantify_scheduler.compatibility_check import check_zhinst_compatibility
+
+check_zhinst_compatibility()